+++ src/main/scala/tbank/academy/scala/fractals/drawer/HitData.scala
@@ -3,7 +3,7 @@ package tbank.academy.scala.fractals.drawer
 case class HitData(
     imageX: Int,
     imageY: Int,
-    red: Int,
-    green: Int,
-    blue: Int
+    red: Double,
+    green: Double,
+    blue: Double
 )
